+++ example/normal_example.dart
@@ -1,32 +1,36 @@
 import 'package:normal/normal.dart';
 
 void main() {
+  final d = Normal(),
+      xs = [for (var i = 0; i < 14; i++) (i - 6) / 2],
+      ps = [for (var i = 1; i < 10; i++) i / 10.0];
+
   // Probability density function.
   print('\nNormal.pdf\n----------\n');
-  for (final x in List.generate(14, (i) => (i - 6) / 2)) {
-    final p = Normal.pdf(x);
+  for (final x in xs) {
+    final p = d.pdf(x);
     print(
         '${x.toString().padLeft(4)} |${' ' * (p * 30).round()}:${p.toStringAsFixed(4)}');
   }
 
   // Cumulative density function.
   print('\nNormal.cdf\n----------\n');
-  for (final x in List.generate(14, (i) => (i - 6) / 2)) {
-    final p = Normal.cdf(x);
+  for (final x in xs) {
+    final p = d.cdf(x);
     print(
         '${x.toString().padLeft(4)} |${' ' * (p * 40).round()}:${p.toStringAsFixed(4)}');
   }
 
   // Quantile function.
   print('\nNormal.inverseCdf\n-----------------\n');
-  for (final p in List.generate(9, (i) => (i + 1) * 0.1)) {
-    final x = Normal.quantile(p);
+  for (final p in ps) {
+    final x = d.quantile(p);
     print('${p.toStringAsFixed(1)}\t${x.toStringAsFixed(4).padLeft(7)}');
   }
 
   // Generate random numbers.
   print('\nNormal.generate\n---------------\n');
-  for (final x in Normal.generate(10, seed: 0)) {
+  for (final x in d.generate(10, seed: 0)) {
     print(x.toStringAsFixed(4).padLeft(7));
   }
 }