+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.7.0
+
+- Housecleaning.
+- Breaking change: switched from abstract to instantiable class.
+- Just use one of the Box-Muller pairs during random number generation for decoupling and simpler code.
+
 ## 0.6.0
 
 - Null safety; updated style.